ENH: Add spatial decomposition capability for regrid weight generation #478
- Also includes support for the SCRIP NetCDF grid format #468
- New command line interface added for chunked regrid weight generation (ocli.py)
- Includes documentation

+++ src/ocgis/util/helpers.py
@@ -1094,7 +1094,7 @@ def create_unique_global_array(arr):
                 req_select_send_size.wait()
                 select_send_size = select_send_size[0]
 
-                u_src = np.zeros(select_send_size, dtype=np_type)
+                u_src = np.zeros(select_send_size.astype(int), dtype=np_type)
                 req = vm.comm.Irecv([u_src, mpi_type], source=o, tag=tag_overlap)
                 req.wait()
